+++ test/integration/src/segments/user/UserSegment.js
@@ -23,7 +23,7 @@ export default class UserSegment extends MapSegment {
         request.get('http://localhost:8000/api/user/' + query.username).end((err, res) => {
           if (res.ok) {
             var payload = JSON.parse(res.text);
-            dispatch(this._createLoadActionObject(queryId, payload));
+            dispatch(this._createSyncLoadActionObject(queryId, payload));
             resolve();
           } else {
             reject(new Error('Unable to fetch user data!'));